+++ tests/test_routing.py
@@ -1,22 +1,23 @@
-"""PyGeoprocessing 1.0 test suite."""
+"""PyGeoprocessing routing test suite."""
+import os
+import logging
+import shutil
 import tempfile
 import unittest
-import shutil
-import os
 
 from osgeo import gdal
 from osgeo import osr
 import numpy
 import numpy.testing
 
-import logging
+import pygeoprocessing
+import pygeoprocessing.routing
 
 LOGGER = logging.getLogger(__name__)
 
 
 class TestRouting(unittest.TestCase):
     """Tests for pygeoprocessing.routing."""
-
     def setUp(self):
         """Create a temporary workspace that's deleted later."""
         self.workspace_dir = tempfile.mkdtemp()
@@ -29,7 +30,6 @@ class TestRouting(unittest.TestCase):
 
     def test_pit_filling(self):
         """PGP.routing: test pitfilling."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
         base_path = os.path.join(self.workspace_dir, 'base.tif')
         dem_array = numpy.zeros((11, 11))
@@ -62,8 +62,6 @@ class TestRouting(unittest.TestCase):
 
     def test_pit_filling_path_band_checking(self):
         """PGP.routing: test pitfilling catches path-band formatting errors."""
-        import pygeoprocessing.routing
-
         with self.assertRaises(ValueError):
             pygeoprocessing.routing.fill_pits(
                 ('invalid path', 1), 'foo')
@@ -74,7 +72,6 @@ class TestRouting(unittest.TestCase):
 
     def test_pit_filling_nodata_int(self):
         """PGP.routing: test pitfilling with nodata value."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
         base_path = os.path.join(self.workspace_dir, 'base.tif')
         dem_array = numpy.zeros((11, 11), dtype=numpy.int32)
@@ -107,8 +104,6 @@ class TestRouting(unittest.TestCase):
 
     def test_flow_dir_d8(self):
         """PGP.routing: test D8 flow."""
-        import pygeoprocessing.routing
-
         driver = gdal.GetDriverByName('GTiff')
         dem_path = os.path.join(self.workspace_dir, 'dem.tif')
         dem_array = numpy.zeros((11, 11))
@@ -154,8 +149,6 @@ class TestRouting(unittest.TestCase):
 
     def test_flow_accum_d8(self):
         """PGP.routing: test D8 flow accum."""
-        import pygeoprocessing.routing
-
         driver = gdal.GetDriverByName('GTiff')
         flow_dir_path = os.path.join(self.workspace_dir, 'flow_dir.tif')
         # this was generated from a pre-calculated plateau drain dem
@@ -215,8 +208,6 @@ class TestRouting(unittest.TestCase):
 
     def test_flow_accum_d8_flow_weights(self):
         """PGP.routing: test D8 flow accum with flow weights."""
-        import pygeoprocessing.routing
-
         driver = gdal.GetDriverByName('GTiff')
         flow_dir_path = os.path.join(self.workspace_dir, 'flow_dir.tif')
         # this was generated from a pre-calculated plateau drain dem
@@ -340,8 +331,6 @@ class TestRouting(unittest.TestCase):
 
     def test_flow_dir_mfd(self):
         """PGP.routing: test multiple flow dir."""
-        import pygeoprocessing.routing
-
         driver = gdal.GetDriverByName('GTiff')
         dem_path = os.path.join(self.workspace_dir, 'dem.tif')
         # this makes a flat raster with a left-to-right central channel
@@ -407,7 +396,6 @@ class TestRouting(unittest.TestCase):
 
     def test_flow_accum_mfd(self):
         """PGP.routing: test flow accumulation for multiple flow."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
 
         n = 11
@@ -474,7 +462,6 @@ class TestRouting(unittest.TestCase):
 
     def test_flow_accum_mfd_with_weights(self):
         """PGP.routing: test flow accum for mfd with weights."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
 
         n = 11
@@ -590,7 +577,6 @@ class TestRouting(unittest.TestCase):
 
     def test_extract_streams_mfd(self):
         """PGP.routing: stream extraction on multiple flow direction."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
 
         n = 11
@@ -646,7 +632,6 @@ class TestRouting(unittest.TestCase):
 
     def test_distance_to_channel_d8(self):
         """PGP.routing: test distance to channel D8."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
         flow_dir_d8_path = os.path.join(self.workspace_dir, 'flow_dir.d8_tif')
 
@@ -733,7 +718,6 @@ class TestRouting(unittest.TestCase):
 
     def test_distance_to_channel_d8_with_weights(self):
         """PGP.routing: test distance to channel D8."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
         flow_dir_d8_path = os.path.join(self.workspace_dir, 'flow_dir.d8_tif')
 
@@ -870,7 +854,6 @@ class TestRouting(unittest.TestCase):
 
     def test_distance_to_channel_mfd(self):
         """PGP.routing: test distance to channel mfd."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
         flow_dir_mfd_path = os.path.join(
             self.workspace_dir, 'flow_dir_mfd.tif')
@@ -993,7 +976,6 @@ class TestRouting(unittest.TestCase):
 
     def test_distance_to_channel_mfd_with_weights(self):
         """PGP.routing: test distance to channel mfd with weights."""
-        import pygeoprocessing.routing
         driver = gdal.GetDriverByName('GTiff')
         flow_dir_mfd_path = os.path.join(
             self.workspace_dir, 'flow_dir_mfd.tif')